refactor: move auth/updates/shared into feature slices (#84)
- src/features/auth/ — AuthPage + authStore (was src/pages/ + src/store/)
- src/features/updates/ — UpdateBanner + useUpdateCheck (was src/hooks/)
- src/features/branches/types.ts — worktree types (was src/types/worktree.ts)
- src/shared/hooks/useTheme + components — moved from src/hooks/ + src/components/
- Delete src/config.ts, src/hooks/useRecentRepos/useRepos (unused)
- Update all imports to @/ alias
