+++ src/containers/AppContainer.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState, useCallback } from 'react'
 import { Terminal } from 'lucide-react'
-import { useAuthStore } from '../store/authStore.ts'
+import { useAuthStore } from '@/features/auth/stores/authStore'
 import { VIEWER_QUERY } from '../lib/github.ts'
 import DashboardPage from '../pages/DashboardPage.tsx'
 import { FeaturesContext, type Feature } from '../lib/features.ts'
