+++ src/features/auth/exports.ts
@@ -1,0 +1,2 @@
+export { useAuthStore } from './stores/authStore'
+export { default as AuthPage } from './components/AuthPage/AuthPage'
