+++ src/features/branches/queries/useBranches.ts
@@ -1,5 +1,5 @@
 import { useQuery } from '@tanstack/react-query'
-import { useAuthStore } from '@/store/authStore'
+import { useAuthStore } from '@/features/auth/stores/authStore'
 import { createClient, BRANCHES_QUERY } from '@/lib/github'
 import type { Branch } from '@/types/github'
