+++ src/features/pull-requests/queries/useCheckContexts.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query'
 import { createClient, PR_CHECK_CONTEXTS_QUERY } from '@/lib/github'
-import { useAuthStore } from '@/store/authStore'
+import { useAuthStore } from '@/features/auth/stores/authStore'
 import type { CheckRunContext, StatusContextItem } from '@/types/github'
 
 interface CheckContextsResult {
