+++ src/features/updates/exports.ts
@@ -1,0 +1,2 @@
+export { useUpdateCheck, isNewer } from './queries/useUpdateCheck'
+export { UpdateBanner } from './components/UpdateBanner/UpdateBanner'
