+++ src/shared/components/Footer/Footer.tsx
